requirements.txt
Django==5.1
djangorestframework==3.15.2
gunicorn==23.0.0